+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Modify module name `ingestion` to `ingest`.
 - Create giganto's communication part as a separate crate. (giganto-client)
 - Move init tracing to giganto-client crate for oplog logging.
+- Fix packet logic in ingest.
 
 ## [0.7.0] - 2023-01-04